policy: remove policymap dependency
Instead, duplicate the PolicyKey type in this package. The key difference
between this type and the PolicyKey type in pkg/maps/policymap, is that
the DestPort field is in host-byte order in the version in pkg/policy.
This avoids pulling in the BPF dependencies into pkg/policy, and also
continues the work to create an intermediate representation of policy
that is not directly tied to pkg/bpf. The duplicated type is renamed to
Key to avoid stuttering in the type name with the package name.

Signed-off by: Ian Vernon <ian@cilium.io>