compute denied identity sets while policy is calculated for endpoint
The proxy requires knowledge of which identities are explicitly denied in order to correctly
implement policy rules imported by users. Previously, this was done by performing another
iteration over the rules to determine whether requirements were not met at L3. This is
an unneeded iteration over the rules, because the iteration is already performed when
policy is being computed for the endpoint. When policy is calculated for the endpoint,
cache the denied identities at ingress and egress for the endpoint inside the Policy
structure to avoid doing unneeded iterations over the rules.

Signed-off by: Ian Vernon <ian@cilium.io>